todo CRUD 전 로그인 확인(authUser), Alert 추가

+++ server/Router/todolist.js
@@ -1,9 +1,18 @@
 const express = require("express");
 const router = express.Router();
 const database = require("../database");
+const jwt = require("jsonwebtoken");
 
 // todo 입력
 router.post("/todoInput", (req, res) => {
+    if(!req.cookies.authUser) {
+        return res.send({success:0,message: '다시 로그인 해주세요'})
+    }
+    const user = jwt.verify(req.cookies.authUser, process.env.JWT_SECRET,(err,token) => {
+        if(err) return null;
+        else return token;
+    })
+    if(user == null) return res.send({ success:0 })
     database.query(
         "INSERT INTO todolist(do_content, do_date, do_isDone) values (?,?,?)", [req.body.do_content, req.body.do_date, req.body.do_isDone],
     function(err, data){
@@ -18,6 +27,14 @@ router.post("/todoInput", (req, res) => {
 
 // todo 수정
 router.post("/todoUpdate", (req, res) => {
+    if(!req.cookies.authUser) {
+        return res.send({success:0,message: '다시 로그인 해주세요'})
+    }
+    const user = jwt.verify(req.cookies.authUser, process.env.JWT_SECRET,(err,token) => {
+        if(err) return null;
+        else return token;
+    })
+    if(user == null) return res.send({ success:0 })
     database.query(
         "UPDATE todolist SET do_content=?, do_updateDate=? WHERE do_id=?", [req.body.do_content, req.body.do_updateDate, req.body.do_id],
     function(err, data){
@@ -32,6 +49,14 @@ router.post("/todoUpdate", (req, res) => {
 
 // todo 체크 0: check x, 1: check o
 router.post("/todoCheck", (req, res) => {
+    if(!req.cookies.authUser) {
+        return res.send({success:0,message: '다시 로그인 해주세요'})
+    }
+    const user = jwt.verify(req.cookies.authUser, process.env.JWT_SECRET,(err,token) => {
+        if(err) return null;
+        else return token;
+    })
+    if(user == null) return res.send({ success:0 })
     database.query(
         "UPDATE todolist SET do_isDone=? WHERE do_id=?", [req.body.do_isDone, req.body.do_id],
     function(err, data){
@@ -55,7 +80,7 @@ router.get('/todos', (req,res) => {
     })
 })
 
-// todo 상세2 사용 x
+// todo 상세2 사용!!
 router.get('/todolist', (req,res) => {
     database.query('SELECT * FROM todolist', (err, result) => {
         if(err) res.send(err);
@@ -67,6 +92,14 @@ router.get('/todolist', (req,res) => {
 
 // todo 삭제
 router.post("/todoDelete", (req, res) => {
+    if(!req.cookies.authUser) {
+        return res.send({success:0,message: '다시 로그인 해주세요'})
+    }
+    const user = jwt.verify(req.cookies.authUser, process.env.JWT_SECRET,(err,token) => {
+        if(err) return null;
+        else return token;
+    })
+    if(user == null) return res.send({ success:0 })
     database.query(
         "DELETE FROM todolist where do_id=?", [req.body.do_id],
     function(err, data){
